requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
Babel==2.13.1
certifi==2024.8.30
flake8==7.1.1
frozenlist==1.4.1
idna==3.8
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
python-dotenv==1.0.1
redis==5.0.8
typing_extensions==4.12.2
yarl==1.9.4
